refactor: Error 코드 패키지 구조 개선
