--- src/main/kotlin/com/wafflestudio/webgam/global/common/exception/Error.kt
@@ -1,5 +1,0 @@
-package com.wafflestudio.webgam.global.common.exception
-
-interface Error {
-    fun code(): Int
-}
